requirements.txt
absl-py==1.4.0
altgraph==0.17.3
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beem-africa==0.1.2
bitstring==4.0.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cycler==0.11.0
daphne==3.0.2
distlib==0.3.6
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.21.7
ecdsa==0.18.0
esptool==4.6.2
filelock==3.12.0
firebase-admin==6.4.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h5py==3.8.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jax==0.4.12
joblib==1.2.0
keras==2.12.0
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
MouseInfo==0.1.3
msgpack==1.0.7
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pefile==2023.2.7
Pillow==10.1.0
platformdirs==3.5.1
proto-plus==1.23.0
protobuf==4.23.2
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.15.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
redis==5.0.1
reedsolo==1.7.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vidstream==0.0.14
virtualenv==20.23.0
Werkzeug==2.3.6
wrapt==1.14.1
zope.interface==6.1
